+++ CHANGES.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+0.0.3 (2020-06-09)
+------------------
+
+- added ability to add a subtitle (placeholder `{SUBTITLE}`) using `-s, --subtitle`,
+  which can be used for a version number
+- added `-i, --print_templates` option to output templates and the supported placeholders
+
+
 0.0.2 (2020-06-07)
 ------------------